+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Update how Node and Yarn are installed inside the application container
 - Allow ISPF reports to be created as ODA or non-ODA
 - Ensure that non-ISPF reports do not get a value set for `is_oda` (for historical reasons, the app relies on non-ISPF reports and activities not having an `is_oda` value set)
+- Validate that the value of `is_oda` is set for ISPF reports
 
 ## Release 139 - 2023-11-14